Customized the look and feel of the site further
Created additional customization to the site to bring the feel more in line with what my portfolio looks like. Trying to establish a sense of branding.
Created post and posts components. Focused on creating the presentation of posts and how they will be shown on the main page. Posts are currently placeholders.